+++ pcsx2/R5900.h
@@ -14,7 +14,7 @@ extern const char* const bios[256];
 }
 
 extern s32 EEsCycle;
-extern u32 EEoCycle;
+extern u64 EEoCycle;
 
 // used for optimization
 union GPR_reg64 {
@@ -260,9 +260,9 @@ extern void cpuClearInt(uint n);
 extern void GoemonPreloadTlb();
 extern void GoemonUnloadTlb(u32 key);
 
-extern void cpuSetNextEvent( u32 startCycle, s32 delta );
+extern void cpuSetNextEvent( u64 startCycle, s32 delta );
 extern void cpuSetNextEventDelta( s32 delta );
-extern int  cpuTestCycle( u32 startCycle, s32 delta );
+extern int  cpuTestCycle( u64 startCycle, s32 delta );
 extern void cpuSetEvent();
 extern int cpuGetCycles(int interrupt);
